CSA-52: edit profile started, registration redux and page added
