+++ src/components/RegistrationForm/RegistrationForm.tsx
@@ -1,4 +1,4 @@
-import { useState } from 'react';
+import { useState, useRef } from 'react';
 import { useForm, SubmitHandler } from 'react-hook-form';
 import { yupResolver } from '@hookform/resolvers/yup';
 
@@ -8,7 +8,7 @@ import Icon from '../../components/Icon';
 import { registerSchema } from '../../schemas/registerSchema';
 import { register } from '../../store/user/thunks';
 import { useAppDispatch } from '../../hooks/redux';
-// import sprite from '../../images/sprite.svg'
+import sprite from '../../images/sprite.svg';
 
 import css from './RegistrationForm.module.css';
 
@@ -17,7 +17,7 @@ interface FormData {
   email: string;
   password: string;
   confirm_password: string;
-  // avatar?: string;
+  avatar?: File | undefined;
 }
 
 const FormRegistration = () => {
@@ -34,105 +34,114 @@ const FormRegistration = () => {
   });
 
   const onSubmit: SubmitHandler<FormData> = (data) => {
-    console.log(data);
+    if (avatarFile) {
+      data.avatar = avatarFile;
+    }
     dispatch(register(data));
     reset();
   };
 
-  // const [avatarFile, setAvatarFile] = useState<File | null>(null);
-  // const fileInputRef = useRef<HTMLInputElement>(null);
+  const [avatarFile, setAvatarFile] = useState<File | null>(null);
+  const fileInputRef = useRef<HTMLInputElement>(null);
 
-  // const handleAvatarChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-  //     const file = event.target.files?.[0];
-  //     setAvatarFile(file || null);
-  // };
+  const handleAvatarChange = (event: React.ChangeEvent<HTMLInputElement>) => {
+    const file = event.target.files?.[0];
+    setAvatarFile(file || null);
+  };
 
-  // const handleFileButtonClick = () => {
-  //     if (fileInputRef.current) {
-  //         fileInputRef.current.click();
-  //     }
-  // };
+  const handleFileButtonClick = () => {
+    if (fileInputRef.current) {
+      fileInputRef.current.click();
+    }
+  };
 
   return (
     <section className={css.form}>
       <h2 className={css.registerTitle}>Sign up</h2>
       <form onSubmit={handleSubmit(onSubmit)} className={css.divFormRegister}>
-        <div className={css.registerGroup}>
-          <Input
-            register={formRegister('name')}
-            placeholder="Name and surname"
-            error={errors.name?.message}
-          />
-          <Input
-            register={formRegister('email')}
-            placeholder="E-mail"
-            error={errors.email?.message}
-          />
-          <Input
-            register={formRegister('password')}
-            placeholder="Password"
-            type={shouldShowPassword ? 'text' : 'password'}
-            error={errors.password?.message}
-            icon={
-              <Icon
-                className={css.eyeIcon}
-                id={shouldShowPassword ? 'eye' : 'eye-closed'}
-                boxStyles={css.iconBox}
-                onClick={() => {
-                  setShouldShowPassword((p) => !p);
-                }}
+        <div className={css.formWrapper}>
+          <div className={css.registerGroup}>
+            <Input
+              register={formRegister('name')}
+              placeholder="Name and surname"
+              error={errors.name?.message}
+            />
+            <Input
+              register={formRegister('email')}
+              placeholder="E-mail"
+              error={errors.email?.message}
+            />
+            <Input
+              register={formRegister('password')}
+              placeholder="Password"
+              type={shouldShowPassword ? 'text' : 'password'}
+              error={errors.password?.message}
+              icon={
+                <Icon
+                  className={css.eyeIcon}
+                  id={shouldShowPassword ? 'eye' : 'eye-closed'}
+                  boxStyles={css.iconBox}
+                  onClick={() => {
+                    setShouldShowPassword((p) => !p);
+                  }}
+                />
+              }
+            />
+            <Input
+              register={formRegister('confirm_password')}
+              placeholder="Confirm password"
+              type={shouldShowPassword ? 'text' : 'password'}
+              error={errors.confirm_password?.message}
+              icon={
+                <Icon
+                  className={css.eyeIcon}
+                  id={shouldShowPassword ? 'eye' : 'eye-closed'}
+                  boxStyles={css.iconBox}
+                  onClick={() => {
+                    setShouldShowPassword((p) => !p);
+                  }}
+                />
+              }
+            />
+          </div>
+          <div className={css.photoBlock}>
+            <p className={css.photoTitle}>Photo(optional)</p>
+            <div className={css.divPhotoThumb}>
+              <input
+                ref={fileInputRef}
+                id="photo"
+                className={css.photoInput}
+                type="file"
+                accept="image/*"
+                onChange={handleAvatarChange}
               />
-            }
-          />
-          <Input
-            register={formRegister('confirm_password')}
-            placeholder="Confirm password"
-            type={shouldShowPassword ? 'text' : 'password'}
-            error={errors.confirm_password?.message}
-            icon={
-              <Icon
-                className={css.eyeIcon}
-                id={shouldShowPassword ? 'eye' : 'eye-closed'}
-                boxStyles={css.iconBox}
-                onClick={() => {
-                  setShouldShowPassword((p) => !p);
-                }}
-              />
-            }
-          />
+              <Icon id="user" className={css.userIcon} />
+            </div>
+            <button
+              type="button"
+              className={css.customFileButton}
+              onClick={handleFileButtonClick}
+            >
+              Upload
+              <span>
+                <svg className={css.uploadIcon}>
+                  <use xlinkHref={`${sprite}#upload`} />
+                </svg>
+              </span>
+            </button>
+          </div>
         </div>
-        {/* <div>
-                    <p className={css.photoTitle}>Photo(optional)</p>
-                    <div className={css.divPhotoThumb}></div>
-                    <input
-                        ref={fileInputRef}
-                        id="photo"
-                        className={css.photoInput}
-                        type="file"
-                        accept="image/*"
-                        onChange={handleAvatarChange}
-                    />
-                    <button
-                        type="button"
-                        className={css.customFileButton}
-                        onClick={handleFileButtonClick}
-                    >Upload
-                        <span>
-                            <svg className={css.uploadIcon}>
-                                <use xlinkHref={`${sprite}#upload`} />
-                            </svg>
-                        </span>
-                    </button>
-                </div> */}
+        <div>
+          <Button
+            type="submit"
+            size="lg"
+            variant="primary"
+            className={css.loginButton}
+          >
+            Submit
+          </Button>
+        </div>
       </form>
-      <Button
-        type="submit"
-        size="lg"
-        variant="primary"
-        className={css.loginButton}
-      >
-        Submit
-      </Button>
     </section>
   );
 };
